[feat] byte version

+++ src/main/java/com/yingying/distributed/hw2/distributedIO/nativeIO/PrintAction.java
@@ -16,10 +16,9 @@ public class PrintAction extends DIOAction {
 
     @Override
     public void close() {
-        for (int item : buffer) {
+        for (byte item : buffer) {
             pw.print(item);
         }
-
         pw.close();
     }
 }
